Make rcutils_find_lastn much more efficient.
In particular, have it go looking from the end, not the
beginning, which will cause it to get out early once it has
found something.  Note that we have to have a separate check
for the 0th element, since we can't do a <= 0 inside of the
for loop with an unsigned variable.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>