Fix remaining leaks in test_string_map.cpp
This change prevents leaks in test_string_map by relying
on OSRF_TESTING_TOOLS_CPP_SCOPE_EXIT to consisdently free
memory even if an exception is thrown.

This also adds a call to:
    set_failing_allocator_is_failing(failing_allocator, true);
...before calling rcutils_string_map_fini() in the test cases
where a failling allocator is used.

Before this change, this was not always the case preventing
the memory from being actually freed.

Fixes #146

Signed-off-by: Thomas Moulard <tmoulard@amazon.com>