Assume WIN32 HINSTANCE is a void *
According to MSDN, `HINSTANCE` is a `HANDLE` is a `PVOID` is a `void *`
everywhere on Windows. If we assume this, we can drop the problematic
Windows.h include from our header and contain it to a single source
file.

I included a sanity check that will break if this assumption is ever
broken.

Signed-off-by: Scott K Logan <logans@cottsay.net>