Don't leak memory on realloc failing. (#232)
When reallocing a memory region, it is possible for the allocation
to fail.  If it does, and you are directly assigning to the
"original" pointer, you'll lose the pointer and hence leak
the memory.  Store the result of the reallocation in a temporary,
do error checking, and then only reassign char_array->buffer
once we are confident it points to real memory.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>