Don't assume errno is set to 0 on success on Windows
Many Windows functions don't set errno or call SetLastError(0) when they
are successful. This particular case is special because the return value
is the same when the output string is truncated as when an error occurs.

However, the output string truncation isn't an error itself, and so
errno is not explicitly.

Signed-off-by: Scott K Logan <logans@cottsay.net>