Clarify duration arg description in logging macros (#359)
The expectation that the `duration` argument to the logging macros will
be an integral value in milliseconds is not explicit at all, and as such
the documentation for this argument should be clarified.

Prior to this change, this information could only be found inside
`RCUTILS_LOG_CONDITION_THROTTLE_BEFORE`, which is inconvenient.

Signed-off-by: Abrar Rahman Protyasha <aprotyas@u.rochester.edu>
(cherry picked from commit 8da2aace17bcecad3bdf0bebafe23298602f50c0)