Optimize the implementation of rcutils_char_array_strncpy. (#367)
It turns out that this function is used a lot in the rcutils
logging code, so it is on a hot-path.  Thus it is worth optimizing.

The optimization itself consists of recognizing that the current
implementation of rcutils_char_array_strncpy is calling strlen
twice; once explicitly at the beginning of the function, and then
again implicitly while calling strncat.  However, it turns out
that we don't need to recalculate the length either time; we
already know that information based on the buffer_length.  Thus
we use the information we already have and speed up the calculation,
particularly for long strings.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>