Optimize index computation in hash_map_find.
That is, avoid a modulus operator (which uses an expensive divide)
and instead use AND.  The downside is that this only works for
power-of-two capacities, which we now enforce in this implementation.

It turns out that we only use the hash_map in one place (in rcl
logging), and that was already a power-of-two so this restriction
should have limited practical effect.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>