Fix a bug in hash_map_get_next_key_and_data. (#375)
What was happening was that we were incorrectly using the
previous_key for finding our last "spot" in the hash_map
for iteration.  That's because of a silly typo.

This wasn't picked up in the tests because the tests also
had a bug; they were reusing the same key pointer for the
previous and the current key, and because of that they
didn't pick up on the bug.  Fix both of these issues, which
fixes iteration over the hash map.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>