Upgrade rcutils to C++17. (#392)
* Upgrade rcutils to C++17.

This upgrade was more involved than just increasing the
CMAKE_CXX_STANDARD version.  In C++17, apparently 'noexcept'
became part of the type system.  This is problematic because
some functions that we use from the underlying libc, like
'stat', actually specify nothrow in their function declaration.
This causes the type system to reject our mock shims, since
the types are no longer totally compatible.

To work around this, introduce a 'remove_noexcept' type into
the system which effectively removes the 'noexcept' specifier
for any function that it uses.  With that in place, everything
here works with C++17.  See the comments in remove_noexcept.hpp
for details on the provenance of that code.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>