Hand-code logging_macros.h
In my opinion, this makes it far more readable for
humans, and thus easier to update/maintain in the future.
This ends up removing ~500 lines of python code, empy templates,
and associated machinery, plus 1500 lines of generated code.
This also has the the nice side effect of making it
so that rosdoc2 will generate the documentation for all of
these macros.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>